requirements.txt
crewai>=0.28.0
playwright>=1.40.0
google-generativeai>=0.3.1
langchain>=0.1.0
langchain-google-genai>=0.0.5
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio>=3.4.3
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
tenacity>=8.2.0 